+++ store/rootmulti/store.go
@@ -484,6 +484,16 @@ func (rs *Store) LastCommitID() types.CommitID {
 	return info.CommitID()
 }
 
+// PausePruning temporarily pauses the pruning of all individual stores which implement
+// the PausablePruner interface.
+func (rs *Store) PausePruning(pause bool) {
+	for _, store := range rs.stores {
+		if pauseable, ok := store.(types.PausablePruner); ok {
+			pauseable.PausePruning(pause)
+		}
+	}
+}
+
 // Commit implements Committer/CommitStore.
 func (rs *Store) Commit() types.CommitID {
 	var previousHeight, version int64
@@ -509,9 +519,17 @@ func (rs *Store) Commit() types.CommitID {
 		rs.logger.Debug("commit header and version mismatch", "header_height", rs.commitHeader.Height, "version", version)
 	}
 
-	cInfo := commitStores(version, rs.stores, rs.removalMap)
-	cInfo.Timestamp = rs.commitHeader.Time
-	rs.lastCommitInfo.Store(cInfo)
+	func() { // ensure unpause
+		// set the committing flag on all stores to block the pruning
+		rs.PausePruning(true)
+		// unset the committing flag on all stores to continue the pruning
+		defer rs.PausePruning(false)
+		cInfo := commitStores(version, rs.stores, rs.removalMap)
+		cInfo.Timestamp = rs.commitHeader.Time
+		rs.lastCommitInfo.Store(cInfo)
+	}()
+
+	cInfo := rs.lastCommitInfo.Load()
 
 	defer rs.flushMetadata(rs.db, version, cInfo)